requirements.txt
Flask==2.3.2
flask-cors==3.1.0
python-dotenv==1.0.0
google-generativeai==0.2.2
os